pyproject.toml
[project]
name = "hve-stem"
version = "0.1.0"
description = "HyperVelocity Stem — control plane for agentic software development"
readme = "README.md"
requires-python = ">=3.12,<3.13"
authors = [{ name = "HVE Stem Team" }]
license = { text = "MIT" }
dependencies = [
    "typer",
    "rich",
    "github-copilot-sdk",
    "mcp>=1.26.0",
    "pyyaml",
    "fastapi",
    "uvicorn[standard]",
    "httpx>=0.27",
]

[project.scripts]
stem = "stem.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stem"]
artifacts = ["src/stem/data/static/**"]

[dependency-groups]
dev = [
    "mypy",
    "pytest",
    "pytest-asyncio",
    "ruff",
    "types-pyyaml",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

